Create a shared tab content wrapper to share styling
