+++ projects/packages/my-jetpack/_inc/components/my-jetpack-tab-panel/styles.module.scss
@@ -1,3 +1,7 @@
 .tab-panel {
-	margin-inline-start: 1rem;
+	margin-inline-start: 2rem;
+}
+
+.tab-content-wrapper {
+	padding-inline-start: 1rem;
 }
